+++ src/main/java/com/hardcode/catalogoprofesores/model/Cursos.java
@@ -19,7 +19,7 @@ public class Cursos implements Serializable{
 	@Id
 	@Column(name="id_curso")
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Long id_curso;
+	private int id_curso;
 	
 	@Column(name="proyecto")
 	private String proyecto;
@@ -45,11 +45,11 @@ public class Cursos implements Serializable{
 		this.nombre_curso = nombre_curso;
 	}
 
-	public Long getId_curso() {
+	public int getId_curso() {
 		return id_curso;
 	}
 
-	public void setId_curso(Long id_curso) {
+	public void setId_curso(int id_curso) {
 		this.id_curso = id_curso;
 	}
